headers: define GS_HAS_DECLARED_PROPERTIES macro
To keep compatibility with compilers prior to Objective-C 2.0 support
(added in the GCC 4.6 series), we are unable to always declare @property
attributes in classes. However, compatibility with dot-syntax
applications require us to define properties if we can.

We already had some checks for the GS_HAS_DECLARED_PROPERTIES macro to
see whether we could or could not declare properties (or whether we
should fall back to manual getter/setter defining). However, it seems
this macro never got defined.

This commit adds this macro, checking whether we possess either GCC >=
4.6 or Clang.